+++ kame/kame/faithd/tcp.c
@@ -1,4 +1,4 @@
-/*	$KAME: tcp.c,v 1.8 2001/11/21 07:40:22 itojun Exp $	*/
+/*	$KAME: tcp.c,v 1.9 2002/05/26 01:17:02 itojun Exp $	*/
 
 /*
  * Copyright (C) 1997 and 1998 WIDE Project.
@@ -92,7 +92,8 @@ sig_child(int sig)
 
 	pid = wait3(&status, WNOHANG, (struct rusage *)0);
 	if (pid > 0 && WEXITSTATUS(status))
-		syslog(LOG_WARNING, "child %d exit status 0x%x", pid, status);
+		syslog(LOG_WARNING, "child %ld exit status 0x%x",
+		    (long)pid, status);
 	exit_success("terminate connection due to child termination");
 }
